Implement isSolvable()
According to the documentation, if the twin of a board is solvable the
board itself is not. So we as we progress through the tree, duplicate
effort on the twin node. As soon as we know if the twin is solvable, we
know that the node we care about is notSolvable.