Improve manhattan()
Before we were not finding the  manhattan distance to the correct item.
Now we are correctly looking at each node and it the number is not what
expect, we find the distance to the node it shoude be in. Unless the
current number is zeo, in which case we skip since we don't care how far
off the zero is.