remove Node arraylists
Instead of carrying around arraylists in the solver class, we only keep
a goal node if the puzzle is solveable. If the goalNode is null, we know
the puzzle is solveable. In order to return the solution, we look at the
predecessor nodes while pushing the nodes into a stack until there is no
more predecessors. Then we return the Stack.