Return the correct number of moves
Previously we were incrementing the number of moves based on how many
times we searched neighbors, regardless if we were adding them to the
tree of nodes. Now when adding a new Node, it will look at the
predecessor (if there is one) and incrememnt the move count based on
that.