Implement PointSet and KdTree
PointSet is bruteforce method of just parsing through all the points to
find what is requested. KdTree puts the data into a BST alternating
which dimension to use as the key in order to partition the data. So as
t depth increases it will swap between using the x and y coordinate.
This commit passes the tests written but it is unclear if it is 100%
correct becuase it is too slow for the grader. The next step is to try
and speed it up.